requirements3.txt
ansible==2.8.6
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
ciscoconfparse==1.4.7
colorama==0.4.1
cryptography==2.8
dnspython==1.16.0
future==0.18.1
idna==2.8
Jinja2==2.10.3
junos-eznc==2.2.1
jxmlease==1.0.1
lxml==4.4.1
MarkupSafe==1.1.1
napalm==2.5.0
napalm-ansible==1.0.0
ncclient==0.6.6
netaddr==0.7.19
netmiko==2.4.2
nxapi-plumbing==0.5.2
paramiko==2.6.0
passlib==1.7.1
pkg-resources==0.0.0
pycparser==2.19
pyeapi==0.8.2
pyIOSXR==0.53
PyNaCl==1.3.0
pynetbox==4.0.8
pyserial==3.4
PyYAML==5.1.2
requests==2.22.0
scp==0.13.2
six==1.12.0
textfsm==1.1.0
urllib3==1.25.6
